+++ R/dof.R
@@ -115,6 +115,11 @@ dof <- degrees_of_freedom
     model <- bayestestR::bayesian_as_frequentist(model)
   }
 
+  # exceptions. logistf, for instance, prints complete summary
+  if (inherits(model, "logistf")) {
+    return(Inf)
+  }
+
   # 1st try
   dof <- try(stats::df.residual(model), silent = TRUE)